Add files via upload
This is my solution to the programming test provided by Omnicom.  A description of the test requirements can be found in the pdf file here.  The provided .csv data files have been zipped up to save space and need to be uncompressed before they are used.
Project is written in Java 8, utilising Spring Core and uses Maven for project management and Jar file build.